New round: carry the 3-player standing over instead of resetting it
Starting a new round wiped the 3-player history the assignment algorithm
balances on, so the outgoing round's standing was forgotten and the same people
could be picked first all over again. The reset now rewrites every retained
member's ExtraThreePlayerTableCount with the quantity the algorithm itself ranks
on -- its post-assignment ratio (threes+1)/(attended+1) -- centred on the
roster's median, rounded to one decimal and clamped to +/-0.5.

That means the field is a double now, and so is every count derived from it:
CountThreePlayerAssignments and the threePlayerCounts dictionaries, the History
page's PlayerStat, and the Member record in the Simulation bench (an int there
would have thrown on any export taken after a round change). Attendance stays an
int, so the attendance tiers and the 1e-9 ratio epsilon are untouched.

Carrying the ratio rather than a yes/no flag drops the threshold constant
altogether. A flag keeps one bit per member: in the new round everyone sits at
attendance 0, so the flagged rank at 2.0 and everyone else ties at exactly 1.0
-- the most under-served member in the club would get no priority over someone
who sat at a 3-table two evenings ago. The median rather than the mean because
never-played members sit at 1.0 and drag a mean far enough up to push most of
the roster negative, which would stop 0 meaning neutral. The +1/+1 smoothing
is kept deliberately: it is what makes a never-played member the least
attractive 3-duty candidate rather than the most.

The decimals stay out of the UI. The new-round preview now also names the
inactive members it is about to delete, and lists the nudged members by name
only. History rounds both the 3-table count and the % to whole numbers, via int
so a small negative head start renders as 0 and not -0; sorting and the
percentage still run on the exact value. Both the History % column and the
member form's correction field gained a bi-question-circle explainer written for
a normal club member.

+++ Simulation/Program.cs
@@ -352,7 +352,7 @@ class Sim
 {
     readonly List<Guid> present;
     readonly Dictionary<Guid, int> att;
-    readonly Dictionary<Guid, int> threeCounts;
+    readonly Dictionary<Guid, double> threeCounts;
     readonly Dictionary<string, int> meetings;
     readonly Dictionary<string, int> coAtt;
     readonly Dictionary<string, int> sameDayMeetings;
@@ -365,7 +365,7 @@ class Sim
     public readonly int ThreeTables;
     public int EligibleCount => eligible.Count;
 
-    public Sim(List<Guid> present, List<Session> history, Dictionary<Guid, int> extraCounts, DateTime day)
+    public Sim(List<Guid> present, List<Session> history, Dictionary<Guid, double> extraCounts, DateTime day)
     {
         this.present = present;
         att = CountAttendance(history, present);
@@ -854,7 +854,7 @@ class Sim
         return c;
     }
 
-    static Dictionary<Guid, int> CountThreePlayerAssignments(List<Session> history, List<Guid> relevant, Dictionary<Guid, int> extra)
+    static Dictionary<Guid, double> CountThreePlayerAssignments(List<Session> history, List<Guid> relevant, Dictionary<Guid, double> extra)
     {
         var c = relevant.ToDictionary(id => id, id => extra.GetValueOrDefault(id, 0));
         foreach (var s in history)
@@ -941,6 +941,8 @@ class Sim
 // ─── Export records (field-subsets of the app's models, deserialized case-insensitively) ───
 
 record ExportData(List<Member> Members, List<Session> Sessions);
-record Member(Guid Id, string Name, string Email, int ExtraThreePlayerTableCount);
+// ExtraThreePlayerTableCount is fractional: the organizer's start-of-round reset writes a decimal
+// head start there, so an int would fail to deserialise any export taken after a round change.
+record Member(Guid Id, string Name, string Email, double ExtraThreePlayerTableCount);
 record Session(Guid Id, DateTime Date, TimeSpan? StartTime, List<Guid> PresentMemberIds, List<Table> Tables, bool IsFinalized, bool ExcludeFromOptimization);
 record Table(int TableNumber, List<Guid> PlayerIds);
